+++ bin/deploy.sh
@@ -50,26 +50,38 @@ function main() {
 
 	# We _want_ to expand the variables on the client side
 	# shellcheck disable=SC2029
-	ssh -o StrictHostKeyChecking=no -v -p "${ssh_port}" "${target_server}" "cd ${remote_path}; mkdir -p \"${short_name}\""
+	if ! ssh -o StrictHostKeyChecking=no -p "${ssh_port}" "${target_server}" "cd ${remote_path}; mkdir -p \"${short_name}\""; then
+		echo "Error: Cannot create ${short_name} directory in ${remote_path}"
+		die 1
+	fi
 
 	echo "Copying ${kit_name} to ${remote_server}:${remote_path}/${short_name}/"
-	scp -o StrictHostKeyChecking=no -v -P "${ssh_port}" "${kit_name}" "${target_server}:${remote_path}/${short_name}/"
+	if ! scp -r -o StrictHostKeyChecking=no -P "${ssh_port}" "${kit_name}" "${target_server}:${remote_path}/${short_name}/"; then
+		echo "Error: Cannot copy ${kit_name} to ${remote_server}:${remote_path}/${short_name}/!"
+		die 1
+	fi
 
 	echo "Copying ${metadata} to ${remote_server}:${remote_path}/${short_name}/"
-	scp -o StrictHostKeyChecking=no -v -P "${ssh_port}" "${metadata}" "${target_server}:${remote_path}/${short_name}/"
+	if ! scp -r -o StrictHostKeyChecking=no -P "${ssh_port}" "${metadata}" "${target_server}:${remote_path}/${short_name}/"; then
+		echo "Error: Unable to copy ${metadata} to ${remote_server}:${remote_path}/${short_name}/"
+		die 1
+	fi
 
 	echo "Linking ${short_name}/${short_name}-${version}.zip to ${short_name}.zip on remote server"
 
 	# We _want_ to expand the variables on the client side
 	# shellcheck disable=SC2029
-	ssh -o StrictHostKeyChecking=no -v -p "${ssh_port}" "${target_server}" \
-		"cd ${remote_path}/ ; ln -sf \"${short_name}\"/\"${short_name}\"-\"${version}\".zip \"${short_name}\".zip"
+	if ! ssh -o StrictHostKeyChecking=no -p "${ssh_port}" "${target_server}" \
+		"cd ${remote_path}/ ; ln -sf \"${short_name}\"/\"${short_name}\"-\"${version}\".zip \"${short_name}\".zip" ; then
+		echo "Error: Unable to link ${short_name}/${short_name}-${version}.zip to ${short_name}.zip"
+		die 1
+	fi
 
 	# Return to the root directory
 	cd "${src_path}" || die 1
 
 	# And clean up
-	rm -rf "${dst_path}"
+	rm -rf "${dst_path}" || echo "Error: Unable to clean up ${dst_path}" && die 1
 }
 
 main "$@"