+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## v2.0.6 - 2021-07-16
+- BUG FIX: Disabled verbose SSH mode and added better error checking/handling for operations
 - BUG FIX: Inform the user that we're using an environment variable to set SSH user/host/port settings
 - BUG FIX: Reduce the number of retries during stack startup for the test containers
 - BUG FIX: Update CHANGELOG.md and Makefile to save the changelog source file